requirements.txt
datasets
flax
jax
matplotlib
numpy
optax
seaborn
torch
torchtext
tqdm
tensorboard
spacy
jaxtyping
tokenizers
#nvidia-cudnn-cu11==8.6.0.163
orbax-checkpoint
orbax-export
tiktoken